AI/news_gpt/requirements.txt
requests
openai
fastapi
uvicorn
python-dotenv
tavily-python